requirements.txt
streamlit==1.42.2
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.164.0
pydantic==2.10.6
pytest==8.3.5
